Added a font renderer api header
